7-05

+++ ordermodel/src/main/java/com/mikleo/ordermodel/Model/Order.java
@@ -3,6 +3,7 @@ package com.mikleo.ordermodel.Model;
 import io.searchbox.annotations.JestId;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 
 @Data
@@ -13,8 +14,8 @@ public class Order {
     private  int user_id;
     private String goodname;
     private int goodnum;
-    private double unitprice;
-    private double totalprice;
+    private BigDecimal unitprice;
+    private BigDecimal totalprice;
     private Date ordertime;
     private int orderstate;
     private Date comfirmtime;
