+++ usermodel/src/main/java/com/mikleo/usermodel/Model/Order.java
@@ -2,6 +2,7 @@ package com.mikleo.usermodel.Model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 
 @Data
@@ -11,8 +12,8 @@ public class Order {
     private  int user_id;
     private String goodname;
     private int goodnum;
-    private double unitprice;
-    private double totalprice;
+    private BigDecimal unitprice;
+    private BigDecimal totalprice;
     private Date ordertime;
     private int orderstate;
     private Date comfirmtime;
